# Conflicts:
#	cmd/system-probe/modules/network_tracer.go
#	cmd/system-probe/modules/network_tracer_test.go
#	pkg/network/encoding/encoding.go
#	pkg/network/encoding/encoding_test.go
#	pkg/network/encoding/format.go
#	pkg/network/encoding/json.go
#	pkg/network/encoding/kafka_test.go
#	pkg/network/encoding/protobuf.go

+++ pkg/security/tests/mount_test.go
@@ -322,7 +322,7 @@ func TestMountSnapshot(t *testing.T) {
 	defer bindMountB.unmount(0)
 
 	mountResolver := test.probe.GetResolvers().MountResolver
-	pid := uint32(utils.Getpid())
+	pid := utils.Getpid()
 
 	mounts, err := kernel.ParseMountInfoFile(int32(pid))
 	if err != nil {
